post-info-modals: Modal now properly shows text when loaded
